Refactor atlasgen into a submodule
