+++ .pre-commit-config.yaml
@@ -23,10 +23,3 @@ repos:
       rev: 24.1.1
       hooks:
           - id: black
-    - repo: https://github.com/pre-commit/mirrors-mypy
-      rev: v1.8.0
-      hooks:
-          - id: mypy
-            additional_dependencies:
-                - types-setuptools
-                - types-requests
