+++ MANIFEST.in
@@ -7,7 +7,6 @@ exclude .codecov.yml
 exclude .pre-commit-config.yaml
 
 graft brainglobe_atlasapi *.py
-graft bg_atlasgen *.py
 
 prune tests
 prune tutorials
