+++ brainglobe_atlasapi/atlas_generation/atlas_scripts/allen_mouse.py
@@ -5,11 +5,11 @@ from pathlib import Path
 from allensdk.api.queries.ontologies_api import OntologiesApi
 from allensdk.api.queries.reference_space_api import ReferenceSpaceApi
 from allensdk.core.reference_space_cache import ReferenceSpaceCache
-from bg_atlasapi import descriptors
 from requests import exceptions
 from tqdm import tqdm
 
-from bg_atlasgen.wrapup import wrapup_atlas_from_data
+from brainglobe_atlasapi import descriptors
+from brainglobe_atlasapi.atlas_generation.wrapup import wrapup_atlas_from_data
 
 
 def create_atlas(working_dir, resolution):
