+++ brainglobe_atlasapi/atlas_generation/atlas_scripts/azba_zfish.py
@@ -17,15 +17,15 @@ from pathlib import Path
 
 import numpy as np
 import tifffile
-from bg_atlasapi import utils
-from bg_atlasapi.structure_tree_util import get_structures_tree
 from rich.progress import track
 
-from bg_atlasgen.mesh_utils import (
+from brainglobe_atlasapi import utils
+from brainglobe_atlasapi.atlas_generation.mesh_utils import (
     Region,
     create_region_mesh,
 )
-from bg_atlasgen.wrapup import wrapup_atlas_from_data
+from brainglobe_atlasapi.atlas_generation.wrapup import wrapup_atlas_from_data
+from brainglobe_atlasapi.structure_tree_util import get_structures_tree
 
 PARALLEL = False  # Disable for debugging mesh creation
