+++ brainglobe_atlasapi/atlas_generation/atlas_scripts/humanatlas.py
@@ -8,18 +8,18 @@ import pandas as pd
 import treelib
 import urllib3
 from allensdk.core.structure_tree import StructureTree
-from bg_atlasapi.structure_tree_util import get_structures_tree
 from brainio import brainio
 from rich.progress import track
 
 # import sys
 # sys.path.append("./")
-from bg_atlasgen.mesh_utils import (
+from brainglobe_atlasapi.atlas_generation.mesh_utils import (
     Region,
     create_region_mesh,
     inspect_meshes_folder,
 )
-from bg_atlasgen.wrapup import wrapup_atlas_from_data
+from brainglobe_atlasapi.atlas_generation.wrapup import wrapup_atlas_from_data
+from brainglobe_atlasapi.structure_tree_util import get_structures_tree
 
 
 def prune_tree(tree):
