+++ brainglobe_atlasapi/atlas_generation/atlas_scripts/template_script.py
@@ -4,7 +4,7 @@ has to be renamed to match the name of the atlas (e.g. allen_mouse.py)
 
 __version__ = "0"  # will be used to set minor version of the atlas
 
-from bg_atlasgen.wrapup import wrapup_atlas_from_data
+from brainglobe_atlasapi.atlas_generation.wrapup import wrapup_atlas_from_data
 
 
 def create_atlas(working_dir, resolution):
