+++ brainglobe_atlasapi/atlas_generation/structure_json_to_csv.py
